+++ python/src/connectors/__init__.py
@@ -1,2 +1,8 @@
 from .azure_chat_completion import RESPONSE_OBJECT_KEY, AzureChatCompletion
 from .open_ai_chat_history import OpenAIChatHistory
+
+__all__ = [
+    "RESPONSE_OBJECT_KEY",
+    "AzureChatCompletion",
+    "OpenAIChatHistory",
+]
