+++ python/src/plugins/semantic_function.py
@@ -7,7 +7,7 @@ from python.src.template_engine.handlebars_prompt_template_handler import (
     HandleBarsPromptTemplateHandler,
 )
 
-from ..orchestration.sk_function import SKFunction
+from .sk_function import SKFunction
 
 
 class SemanticFunction(SKFunction):
