+++ python/src/template_engine/__init__.py
@@ -1,1 +1,3 @@
 from .handlebars_prompt_template_handler import HandleBarsPromptTemplateHandler
+
+__all__ = ["HandleBarsPromptTemplateHandler"]
